+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+  * Fixed waypoint.next and .previous causing loops when two opposite-direction lanes follow each other in the map.
+
 ## CARLA 0.9.15
 
   * Added Digital Twins feature version 0.1. Now you can create your own map based on OpenStreetMaps